Entity class  set up

+++ Vec2.cpp
@@ -3,7 +3,7 @@
 
     Vec2::Vec2(){}
 
-    Vec2::Vec2(int x, int y): x(x), y(y){}
+    Vec2::Vec2(double x, double y): x(x), y(y){}
 
     bool Vec2::operator == (const Vec2 & v2) const
     {
@@ -23,11 +23,11 @@
     {
         return Vec2(x - v2.x, y - v2.y);
     }
-    Vec2 Vec2::operator * (const float val) const
+    Vec2 Vec2::operator * (const double val) const
     {
         return Vec2(x * val, y * val);
     }
-    Vec2 Vec2::operator / (const float val) const
+    Vec2 Vec2::operator / (const double val) const
     {
         return Vec2(x / val, y / val);
     }
@@ -42,40 +42,40 @@
         x -= v2.x;
         y -= v2.y;
     }
-    void Vec2::operator *= (const float val)
+    void Vec2::operator *= (const double val)
     {
         x *= val;
         y *= val;
     }
-    void Vec2::operator /= (const float val)
+    void Vec2::operator /= (const double val)
     {
         x /= val;
         y /= val;
     }
 
-    float Vec2::dist(const Vec2 & v2) const
+    double Vec2::dist(const Vec2 & v2) const
     {
         float dx = v2.x - x;
         float dy = v2.y - y;
         return sqrt(dx * dx + dy * dy);
     }
 
-    float Vec2::getX() const
+    double Vec2::getX() const
     {
         return x;
     }
 
-    float Vec2::getY() const
+    double Vec2::getY() const
     {
         return y;
     }
 
-    void Vec2::setX(const float num)
+    void Vec2::setX(const double num)
     {
         x = num;
     }
 
-    void Vec2::setY(const float num)
+    void Vec2::setY(const double num)
     {
         y = num;
     }
